feat: Accessibility Notification/Parsing improvements (#11)
* Retrieve user's text selection (debounce 300ms)
* Retrieve user's input (debounce 300ms)
* Create a `ScreenResult` model which own all extracted datas
* Remove observers on app deactivated
* Improve debugger
* Code cleaning
* Removed hard-coded value
* Create configs for constants
* Refactorize the external element's check
