+++ macos/Onit/Accessibility/AccessibilityParsedElements.swift
@@ -1,0 +1,19 @@
+//
+//  AccessibilityParsedElements.swift
+//  Onit
+//
+//  Created by Kévin Naudin on 23/01/2025.
+//
+
+struct AccessibilityParsedElements {
+    static let applicationName = "applicationName"
+    static let applicationTitle = "applicationTitle"
+    static let elapsedTime = "elapsedTime"
+    
+    static let screen = "screen"
+    
+    struct Xcode {
+        static let editor = "editor"
+        static let console = "console"
+    }
+}
